requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=2.0.0
googlemaps>=4.10.0
geopy>=2.4.0
python-dotenv>=1.0.0
folium>=0.14.0
streamlit-folium>=0.15.0

